RecyclerViewItemModel works but recompose is buggy.
This approach reduces the RecyclerView's ViewModel down to simple uniqueIds that are used to start a coroutine job that will gather the complete item View Model per view that is actually on screen.

This is buggy because there are problems not worth fixing. It seems that a composable is not able to be later updated by a flow? There are concepts like state hoisting that seem to almost allow this but I don't think repositories should be passed to composables. Back to regular ol' views.